requirements.txt
langchain
langchain-openai
langchain-community
langgraph
chromadb
streamlit
pandas
tiktoken
python-dotenv
Pillow
